refactor: split userRunHistory infra
